Initial development for base frame of app. (#1)
* Add menubar and gitignore for node modules

* Testing electron build for weather widget

* getWeather works with hard coded token and city id, but cannot prompt user to set env var yet.

* Clean up index.js and index.html

* Add editor config.

* Update main.js. Building from electron scratch. Remove menubar.

* Electron app creates tray icon. weather displayed on hover.
Window is hidden and shown on icon click.

* Start app minimized. Add date to tool tip to help debug if weather is being updated on hover. Add on mouse-enter listener to icon.

* Remvoing useless file that was left over in last clean

* Readding index.html and removing index.js

* Setting the window to have no frame so that a windows user does not close it by pressing 'x'

* Moving getWeather into an "on finish loading" event so that we can make
a fresh API call when we open app is done loading initially. We can then
pass this to the HMTL file which can now handle the input that being
passed to it.
Modifying index.html to now have an empty paragraph and a placeholder
refresh button. The button is not wired up and the paragraph gets
populated by the script tag that reads input from the electron event.

* Remove undeeded boostrap cdn